fix: 파티 조회 시 radius 속성 추가
- radius 속성이 누락됨
Ref: #7
Related to: #8

+++ src/main/java/com/wap/app2/gachitayo/dto/response/PartyResponseDto.java
@@ -14,7 +14,7 @@ public record PartyResponseDto(
         @JsonProperty("party_stopovers")
         List<StopoverDto> stopovers,
         @JsonProperty("party_radius")
-        Integer radius,
+        Double radius,
         @JsonProperty("party_max_person")
         Integer maxPerson,
         @JsonProperty("party_option")
